+++ campaign-service/src/main/java/com/oem/evwarranty/model/Campaign.java
@@ -17,7 +17,7 @@ public class Campaign {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "campaign_id")
-    private Integer id;
+    private Long id;
 
     @Column(length = 32, nullable = false)
     private String code;
@@ -50,8 +50,8 @@ public class Campaign {
     private List<Notification> notifications = new ArrayList<>();
 
     // getters/setters
-    public Integer getId() { return id; }
-    public void setId(Integer id) { this.id = id; }
+    public Long getId() { return id; }
+    public void setId(Long id) { this.id = id; }
 
     public String getCode() { return code; }
     public void setCode(String code) { this.code = code; }
