+++ campaign-service/src/main/java/com/oem/evwarranty/repository/NotificationRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface NotificationRepository extends JpaRepository<Notification, Integer> {
-    Page<Notification> findByCampaignId(Integer campaignId, Pageable pageable);
-    Page<Notification> findByAffectedId(Integer affectedId, Pageable pageable);
+public interface NotificationRepository extends JpaRepository<Notification, Long> {
+    Page<Notification> findByCampaignId(Long campaignId, Pageable pageable);
+    Page<Notification> findByAffectedId(Long affectedId, Pageable pageable);
 }
